+++ src/test/java/io/github/astrapi69/crypto/key/reader/PemObjectReaderTest.java
@@ -25,14 +25,21 @@
 package io.github.astrapi69.crypto.key.reader;
 
 import static org.testng.AssertJUnit.assertEquals;
+import static org.testng.AssertJUnit.assertFalse;
 import static org.testng.AssertJUnit.assertNotNull;
 import static org.testng.AssertJUnit.assertNull;
 import static org.testng.AssertJUnit.assertTrue;
 
 import java.io.File;
 import java.io.IOException;
+import java.security.NoSuchAlgorithmException;
+import java.security.NoSuchProviderException;
 import java.security.PrivateKey;
+import java.security.Security;
+import java.security.spec.InvalidKeySpecException;
+import java.util.Optional;
 
+import org.bouncycastle.jce.provider.BouncyCastleProvider;
 import org.bouncycastle.util.io.pem.PemObject;
 import org.meanbean.test.BeanTester;
 import org.testng.annotations.BeforeMethod;
@@ -58,6 +65,7 @@ public class PemObjectReaderTest
 	@BeforeMethod
 	protected void setUp()
 	{
+		Security.addProvider(new BouncyCastleProvider());
 		pemDir = new File(PathFinder.getSrcTestResourcesDir(), "pem");
 		derDir = new File(PathFinder.getSrcTestResourcesDir(), "der");
 		privateKeyDerFile = new File(derDir, "private.der");
@@ -195,17 +203,38 @@ public class PemObjectReaderTest
 	{
 		File privatekeyPemDir;
 		File privatekeyPemFile;
-		PrivateKey privateKey;
+		Optional<PrivateKey> actual;
 		// new scenario...
 		privatekeyPemDir = new File(PathFinder.getSrcTestResourcesDir(), "pem");
 		privatekeyPemFile = new File(privatekeyPemDir, "id_rsa");
 
-		privateKey = PemObjectReader.readPemPrivateKey(privatekeyPemFile, "secret");
-		assertNotNull(privateKey);
+		actual = PemObjectReader.readPemPrivateKey(privatekeyPemFile, "secret");
+		assertTrue(actual.isPresent());
 		// new scenario...
 		privatekeyPemFile = new File(privatekeyPemDir, "test.key");
 
-		privateKey = PemObjectReader.readPemPrivateKey(privatekeyPemFile, "bosco");
+		actual = PemObjectReader.readPemPrivateKey(privatekeyPemFile, "bosco");
+		assertTrue(actual.isPresent());
+	}
+
+	/**
+	 * Test method for {@link PemObjectReader#readPemPrivateKey(File)}
+	 *
+	 * @throws IOException
+	 *             Signals that an I/O exception has occurred.
+	 */
+	@Test
+	public void testReadPemPrivateKeyWithoutPassword() throws IOException, NoSuchAlgorithmException,
+		InvalidKeySpecException, NoSuchProviderException
+	{
+		File privatekeyPemDir;
+		File privatekeyPemFile;
+		PrivateKey privateKey;
+		// new scenario...
+		privatekeyPemDir = new File(PathFinder.getSrcTestResourcesDir(), "pem");
+		privatekeyPemFile = new File(privatekeyPemDir, "private.pem");
+
+		privateKey = PemObjectReader.readPemPrivateKey(privatekeyPemFile);
 		assertNotNull(privateKey);
 	}
 
@@ -240,7 +269,7 @@ public class PemObjectReaderTest
 	@Test
 	public void testToPemFormatFile() throws IOException
 	{
-		String actual;
+		Optional<String> actual;
 		File privatekeyPemDir;
 		File privatekeyPemFile;